refactor: optimized package structure
